+++ datacamp_downloader/datacamp.py
@@ -1,0 +1,89 @@
+from bs4 import BeautifulSoup
+from helper import Logger
+from constants import *
+
+
+class Datacamp:
+    def __init__(self, session) -> None:
+
+        self.username = None
+        self.password = None
+        self.token = None
+        self.session = session
+        self.has_active_subscription = False
+        self.loggedin = False
+        self.login_data = None
+
+    def login(self, username, password):
+        if username == self.username and self.password == password and self.loggedin:
+            Logger.info("Already logged in!")
+            return
+
+        self.session.restart()
+        self.username = username
+        self.password = password
+
+        req = self.session.get(LOGIN_URL)
+        if not req or req.status_code != 200 or not req.text:
+            Logger.error("Cannot access datacamp website!")
+            return
+        soup = BeautifulSoup(req.text, "html.parser")
+        authenticity_token = soup.find("input", {"name": "authenticity_token"}).get(
+            "value"
+        )
+        if not authenticity_token:
+            Logger.error("Cannot find authenticity_token attribute!")
+            return
+
+        login_req = self.session.post(
+            LOGIN_URL,
+            data=LOGIN_DATA.format(
+                email=username, password=password, authenticity_token=authenticity_token
+            ).encode("utf-8"),
+        )
+        if (
+            not login_req
+            or login_req.status_code != 200
+            or "/users/sign_up" in login_req.text
+        ):
+            Logger.error("Incorrent username/password")
+            return
+
+        self._set_profile()
+
+    def set_token(self, token):
+        if self.token == token and self.loggedin:
+            Logger.info("Already logged in!")
+            return
+
+        self.session.restart()
+        self.token = token
+        cookie = {
+            "name": "_dct",
+            "value": token,
+            "domain": ".datacamp.com",
+            "secure": True,
+        }
+        self.session.add_cookie(cookie)
+        self._set_profile()
+
+    def _set_profile(self):
+        page = self.session.get(LOGIN_DETAILS_URL)
+        try:
+            data = page.json()
+        except:
+            Logger.error("Please provide a valid token!")
+            return
+
+        Logger.info("Hi, " + data["first_name"])
+
+        if data["has_active_subscription"]:
+            Logger.info("Active subscription found")
+        else:
+            Logger.warning("No active subscription found")
+
+        self.loggedin = True
+        self.login_data = data
+        self.has_active_subscription = data["has_active_subscription"]
+
+        self.session.save()
